Added index-building scripts

+++ scripts/make_grch38.sh
@@ -1,8 +1,11 @@
 #!/bin/sh
 
 #
-# Downloads sequence for the MM10 version of M. musculus (mouse) from
-# UCSC.
+# Downloads sequence for the GRCh38 release 81 version of H. spiens (human) from
+# Ensembl.
+#
+# Note that Ensembl's GRCh81 build has three categories of compressed fasta
+# files:
 #
 # The base files, named ??.fa.gz
 #
@@ -12,8 +15,8 @@
 # variable below.
 #
 
-UCSC_MM10_BASE=http://hgdownload.cse.ucsc.edu/goldenPath/mm10/bigZips
-F=chromFa.tar.gz
+ENSEMBL_RELEASE=81
+ENSEMBL_GRCh38_BASE=ftp://ftp.ensembl.org/pub/release-${ENSEMBL_RELEASE}/fasta/homo_sapiens/dna
 
 get() {
 	file=$1
@@ -41,9 +44,13 @@ if [ ! -x "$HISAT2_BUILD_EXE" ] ; then
 fi
 
 rm -f genome.fa
-get ${UCSC_MM10_BASE}/$F || (echo "Error getting $F" && exit 1)
-tar xvzfO $F > genome.fa || (echo "Error unzipping $F" && exit 1)
-rm $F
+F=Homo_sapiens.GRCh38.dna.primary_assembly.fa
+if [ ! -f $F ] ; then
+	get ${ENSEMBL_GRCh38_BASE}/$F.gz || (echo "Error getting $F" && exit 1)
+	gunzip $F.gz || (echo "Error unzipping $F" && exit 1)
+	awk '{if($1 ~ /^>/) {print $1} else {print}}' $F > genome.fa
+	rm $F
+fi
 
 CMD="${HISAT2_BUILD_EXE} genome.fa genome"
 echo Running $CMD
